moved the fetch all posts function to app level component and passing it down to childs. Feat: added location and author username filtering to the search bar
