+++ src/components/Home.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import GetAllPosts from "./GetAllPosts";
 
-function Home( {isLoggedIn}) {
+function Home({ isLoggedIn, posts }) {
   return (
     <div>
       <h1>Home</h1>
       <br />
-      <GetAllPosts isLoggedIn={isLoggedIn} />
+      <GetAllPosts isLoggedIn={isLoggedIn} posts={posts} />
     </div>
   );
 }
